+++ libs/data/project.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "name": "data",
+  "projectType": "library",
+  "sourceRoot": "libs/data/src",
+  "tags": ["type:data"],
+  "targets": {
+    "lint": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "eslint libs/data/src",
+        "cwd": "."
+      }
+    }
+  }
+}
