+++ libs/data/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./lib/product.types";
+export * from "./lib/products.api";
