+++ libs/data/src/lib/product.types.ts
