+++ libs/hooks/src/index.ts
@@ -1,9 +1,6 @@
-export * from "./lib/product.types";
-export * from "./lib/products.api";
-export * from "./lib/products.queries";
+export * from "./lib/useProductsQueries";
 export * from "./lib/useDebounce";
 export * from "./lib/useLocalStorage";
 export * from "./lib/useSidebar";
-export * from "./lib/SidebarProvider";
-export * from "./lib/cartStore";
-export * from "./lib/themeStore";
+export * from "./lib/useCartStore";
+export * from "./lib/useThemeStore";
