--- libs/hooks/src/lib/SidebarContext.tsx
@@ -1,11 +1,0 @@
-import { createContext } from "react";
-
-export interface SidebarContextType {
-  isOpen: boolean;
-  openSidebar: () => void;
-  closeSidebar: () => void;
-  toggleSidebar: () => void;
-}
-
-export const SidebarContext =
-  createContext<SidebarContextType | null>(null);
