--- libs/hooks/src/lib/useSidebar.ts
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { SidebarContext } from "./SidebarContext";
-
-export const useSidebar = () => {
-  const ctx = useContext(SidebarContext);
-  if (!ctx) {
-    throw new Error("useSidebar must be used inside SidebarProvider");
-  }
-  return ctx;
-};
