+++ libs/hooks/src/lib/useSidebar.tsx
@@ -1,0 +1,35 @@
+import { createContext, useContext, type ReactNode } from "react";
+import { useLocalStorage } from "./useLocalStorage";
+
+interface SidebarContextType {
+  isOpen: boolean;
+  openSidebar: () => void;
+  closeSidebar: () => void;
+  toggleSidebar: () => void;
+}
+
+const SidebarContext = createContext<SidebarContextType | null>(null);
+
+export const SidebarProvider = ({ children }: { children: ReactNode }) => {
+  const [isOpen, setOpen] = useLocalStorage("sidebar-open", false);
+
+  const openSidebar = () => setOpen(true);
+  const closeSidebar = () => setOpen(false);
+  const toggleSidebar = () => setOpen((value) => !value);
+
+  return (
+    <SidebarContext.Provider
+      value={{ isOpen, openSidebar, closeSidebar, toggleSidebar }}
+    >
+      {children}
+    </SidebarContext.Provider>
+  );
+};
+
+export const useSidebar = () => {
+  const ctx = useContext(SidebarContext);
+  if (!ctx) {
+    throw new Error("useSidebar must be used inside SidebarProvider");
+  }
+  return ctx;
+};
