+++ tsconfig.base.json
@@ -18,6 +18,8 @@
       "@productexplorer/ui/*": ["libs/ui/src/*"],
       "@productexplorer/hooks": ["libs/hooks/src/index.ts"],
       "@productexplorer/hooks/*": ["libs/hooks/src/*"],
+      "@productexplorer/data": ["libs/data/src/index.ts"],
+      "@productexplorer/data/*": ["libs/data/src/*"],
       "@productexplorer/i18n": ["libs/i18n/src/index.ts"],
       "@productexplorer/i18n/*": ["libs/i18n/src/*"]
     }
